+++ src/main/scala/net/hcoop/smallory/scalaprs/Alert/GeneralAlert.scala
@@ -12,11 +12,11 @@ import net.hcoop.smallory.scalaprs._
 class GeneralAlert (dataField: String) extends Alert with Serializable {
   var comparison: String = "<"
   var compareTo: String = ""
-  var limit: Float = 0.0f
+  var limit: Double = 0.0d
   var message: String = s"Field $dataField triggered warning."
   val models: Vector[String] = Vector(dataField)
 
-  override def value(models: Map[String, Model]): Float = {
+  override def value(models: Map[String, Model]): Double = {
     comparison match {
       case "<" | "<=" | "le" | "lt" => return models(dataField).min()
       case ">" | ">=" | "ge" | "gt" => return models(dataField).max()
@@ -31,12 +31,12 @@ object GeneralAlert {
     val ta = new GeneralAlert(setup(0))
     setup.length match {
       case 2 => try {
-        ta.limit = setup(1).toFloat
+        ta.limit = setup(1).toDouble
       } catch {case e: Exception => logDataError(
         "Reading user file, got "+e.toString) }
       case 3 => try { 
         ta.comparison = setup(1)
-        ta.limit = setup(2).toFloat
+        ta.limit = setup(2).toDouble
       } catch {case e: Exception => logDataError(
         "Reading user file, got "+e.toString) }
       case _ => ;
