+++ src/main/scala/net/hcoop/smallory/scalaprs/Alert/TemperatureAlert.scala
@@ -9,11 +9,11 @@ import scala.Exception
 class TemperatureAlert extends Alert with Serializable {
   var comparison: String = "<"
   var compareTo: String = ""
-  var limit: Float = 32.0f
+  var limit: Double = 32.0d
   val message: String = "Freezing temperature rather likely."
   val models: Vector[String] = Vector("temperature")
 
-  override def value(models: Map[String, Model]): Float = {
+  override def value(models: Map[String, Model]): Double = {
     return models("temperature").min()
   }
 }
@@ -23,11 +23,11 @@ object TemperatureAlert {
     val ta = new TemperatureAlert()
     setup.length match {
       case 2 => try {
-        ta.limit = setup(1).toFloat
+        ta.limit = setup(1).toDouble
       } catch {case e: Exception => {} }
       case 3 => try { 
         ta.comparison = setup(1)
-        ta.limit = setup(2).toFloat
+        ta.limit = setup(2).toDouble
       } catch {case e: Exception => {} }
       case _ => ;
     }
